Update to v1.4.4.2
1. Fixed inconsistency of chromedriver versions with Google Chrome due to incomplete update. https://github.com/rzc0d3r/ESET-KeyGen/issues/170

When an update is downloaded, Google Chrome changes the manifest settings to indicate the new version. Chromedriver is downloaded relative to the manifest settings. But unless the user has closed browser sessions, the new version of the browser is the new_chrome.exe file, and the current chrome.exe is the past version. Thus the new chromedriver refers to the old chrome.exe although in fact there is a new and appropriate new_chrome.exe in the folder. That's why the error "version mismatch" appeared.

There were 2 solutions:
- When you completely restart all Google Chrome processes, it replaces chrome.exe with new_chrome.exe. This completes the update process completely and the project initializes chromedriver correctly.

- Detect that the Google Chrome update process is not completely finished and refer to new_chrome.exe instead of chrome.exe (This was implemented in this version).

2. The pyperclip module has been removed and __TempMailAPI__ initialization has been slightly changed, it became slower but more stable!

3. Added command line argument ```--try-auto-cloudflare``` (https://github.com/rzc0d3r/ESET-KeyGen/blob/main/wiki/CommandLineArguments.md)